+++ NEWS.md
@@ -32,6 +32,9 @@ Version 0.13b0 - dev
 - NEW: `UCS` and `OCS` uses `Matrix44`for transformations
 - NEW: `Hatch.set_pattern_scale()` to set scaling of pattern definition
 - NEW: `Hatch.set_pattern_angle()` to set rotation angle of pattern definition
+- NEW: `Hatch.paths.arc_edges_to_ellipse_edges()` convert bulges and arcs to ellipse edges
+- NEW: `Hatch.paths.ellipse_edges_to_spline_edges()` convert ellipse edges to spline_edges
+- NEW: `Hatch.paths.all_to_spline_edges()` convert all curves to spline_edges
 - NEW: `BSpline.from_fit_points()` returns a B-Spline from fit points
 - NEW: `BSpline.from_arc()` returns a B-Spline approximation for an arc defined by `ConstructionArc()`
 - NEW: `BSpline.from_ellipse()` returns a B-Spline approximation for an ellipse defined by `ConstructionEllipse()`